requirements.txt
flask
pillow
numpy
onnxruntime
onnx